Move xar applying usage in separate class

+++ Autoupdate/SUBinaryDeltaApply.m
@@ -8,33 +8,23 @@
 
 #import "SUBinaryDeltaApply.h"
 #import "SUBinaryDeltaCommon.h"
+#import "SPUXarDeltaArchive.h"
 #include <CommonCrypto/CommonDigest.h>
 #import <Foundation/Foundation.h>
 #include "bspatch.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include <xar/xar.h>
-#include <Availability.h>
 
 
 #include "AppKitPrevention.h"
 
-#if __MAC_OS_X_VERSION_MAX_ALLOWED >= 120000
-    #define HAS_XAR_GET_SAFE_PATH 1
-#else
-    #define HAS_XAR_GET_SAFE_PATH 0
-#endif
-
-#if HAS_XAR_GET_SAFE_PATH
-// This is preferred over xar_get_path (which is deprecated) when it's available
-// Don't use OS availability for this API
-extern char *xar_get_safe_path(xar_file_t f) __attribute__((weak_import));
-#endif
-
-static BOOL applyBinaryDeltaToFile(xar_t x, xar_file_t file, NSString *sourceFilePath, NSString *destinationFilePath)
+static BOOL applyBinaryDeltaToFile(SPUXarDeltaArchive *archive, const void *item, NSString *sourceFilePath, NSString *destinationFilePath)
 {
     NSString *patchFile = temporaryFilename(@"apply-binary-delta");
-    xar_extract_tofile(x, file, [patchFile fileSystemRepresentation]);
+    if (![archive extractItem:item destination:patchFile]) {
+        return NO;
+    }
+    
     const char *argv[] = {"/usr/bin/bspatch", [sourceFilePath fileSystemRepresentation], [destinationFilePath fileSystemRepresentation], [patchFile fileSystemRepresentation]};
     BOOL success = (bspatch(4, argv) == 0);
     unlink([patchFile fileSystemRepresentation]);
@@ -43,8 +33,8 @@ static BOOL applyBinaryDeltaToFile(xar_t x, xar_file_t file, NSString *sourceFil
 
 BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFile, BOOL verbose, void (^progressCallback)(double progress), NSError *__autoreleasing *error)
 {
-    xar_t x = xar_open([patchFile fileSystemRepresentation], READ);
-    if (!x) {
+    id<SPUDeltaArchiveProtocol> archive = [[SPUXarDeltaArchive alloc] initWithPatchFileForReading:patchFile];
+    if (archive == nil) {
         if (error != NULL) {
             *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileReadUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unable to open %@. Giving up.", patchFile] }];
         }
@@ -52,58 +42,16 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
     }
 
     SUBinaryDeltaMajorVersion majorDiffVersion = FIRST_DELTA_DIFF_MAJOR_VERSION;
-    int minorDiffVersion = 0;
+    uint16_t minorDiffVersion = 0;
 
     NSString *expectedBeforeHash = nil;
     NSString *expectedAfterHash = nil;
 
     progressCallback(0/6.0);
-
-    xar_subdoc_t subdoc;
-    for (subdoc = xar_subdoc_first(x); subdoc; subdoc = xar_subdoc_next(subdoc)) {
-        if (strcmp(xar_subdoc_name(subdoc), BINARY_DELTA_ATTRIBUTES_KEY) == 0) {
-
-            // available in version 2.0 or later
-            {
-                const char *value = NULL;
-                xar_subdoc_prop_get(subdoc, MAJOR_DIFF_VERSION_KEY, &value);
-                if (value != NULL) {
-                    majorDiffVersion = (SUBinaryDeltaMajorVersion)[@(value) intValue];
-                }
-            }
-
-            // available in version 2.0 or later
-            {
-                const char *value = NULL;
-                xar_subdoc_prop_get(subdoc, MINOR_DIFF_VERSION_KEY, &value);
-                if (value != NULL) {
-                    minorDiffVersion = [@(value) intValue];
-                }
-            }
-
-            // available in version 2.0 or later
-            {
-                const char *value = NULL;
-                xar_subdoc_prop_get(subdoc, BEFORE_TREE_SHA1_KEY, &value);
-                if (value != NULL) {
-                    expectedBeforeHash = @(value);
-                }
-            }
-
-            // available in version 2.0 or later
-            {
-                const char *value = NULL;
-                xar_subdoc_prop_get(subdoc, AFTER_TREE_SHA1_KEY, &value);
-                if (value != NULL) {
-                    expectedAfterHash = @(value);
-                }
-            }
-        }
-    }
+    
+    [archive getMajorDeltaVersion:&majorDiffVersion minorDeltaVersion:&minorDiffVersion beforeTreeHash:&expectedBeforeHash afterTreeHash:&expectedAfterHash];
 
     if (majorDiffVersion < FIRST_DELTA_DIFF_MAJOR_VERSION) {
-        xar_close(x);
-        
         if (error != NULL) {
             *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileReadCorruptFileError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unable to identify diff-version %u in delta.  Giving up.", majorDiffVersion] }];
         }
@@ -111,8 +59,6 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
     }
     
     if (majorDiffVersion < FIRST_SUPPORTED_DELTA_MAJOR_VERSION) {
-        xar_close(x);
-        
         if (error != NULL) {
             *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileReadCorruptFileError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Applying version %u patches is no longer supported.", majorDiffVersion] }];
         }
@@ -120,31 +66,23 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
     }
 
     if (majorDiffVersion > LATEST_DELTA_DIFF_MAJOR_VERSION) {
-        xar_close(x);
-        
         if (error != NULL) {
             *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileReadUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"A later version is needed to apply this patch (on major version %u, but patch requests version %u).", LATEST_DELTA_DIFF_MAJOR_VERSION, majorDiffVersion] }];
         }
         return NO;
     }
     
-#if HAS_XAR_GET_SAFE_PATH
     // Reject patches that did not generate valid hierarchical xar container paths
     // These will not succeed to patch using recent versions of BinaryDelta
-    if (majorDiffVersion == SUBinaryDeltaMajorVersion2 && minorDiffVersion < 3) {
-        xar_close(x);
-        
+    if ([[archive class] supportsSafeExtraction] && majorDiffVersion == SUBinaryDeltaMajorVersion2 && minorDiffVersion < 3) {
         if (error != NULL) {
             *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileReadCorruptFileError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"This patch version (%u.%u) is too old and potentially unsafe to apply. Please re-generate the patch using the latest version of BinaryDelta or generate_appcast. New version %u.%u patches will still be compatible with older versions of Sparkle.", majorDiffVersion, minorDiffVersion, majorDiffVersion, latestMinorVersionForMajorVersion(majorDiffVersion)] }];
         }
         
         return NO;
     }
-#endif
 
     if (expectedBeforeHash == nil || expectedAfterHash == nil) {
-        xar_close(x);
-        
         if (error != NULL) {
             *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileReadCorruptFileError userInfo:@{ NSLocalizedDescriptionKey: @"Unable to find before-sha1 or after-sha1 metadata in delta.  Giving up." }];
         }
@@ -160,8 +98,6 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
 
     NSString *beforeHash = hashOfTreeWithVersion(source, majorDiffVersion);
     if (!beforeHash) {
-        xar_close(x);
-        
         if (verbose) {
             fprintf(stderr, "\n");
         }
@@ -172,8 +108,6 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
     }
 
     if (![beforeHash isEqualToString:expectedBeforeHash]) {
-        xar_close(x);
-        
         if (verbose) {
             fprintf(stderr, "\n");
         }
@@ -190,8 +124,6 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
     progressCallback(2/6.0);
 
     if (!removeTree(destination)) {
-        xar_close(x);
-        
         if (verbose) {
             fprintf(stderr, "\n");
         }
@@ -204,8 +136,6 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
     progressCallback(3/6.0);
 
     if (!copyTree(source, destination)) {
-        xar_close(x);
-        
         if (verbose) {
             fprintf(stderr, "\n");
         }
@@ -221,36 +151,18 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
         fprintf(stderr, "\nPatching...");
     }
     NSFileManager *fileManager = [[NSFileManager alloc] init];
-    xar_file_t file;
-    xar_iter_t iter = xar_iter_new();
-    for (file = xar_file_first(x, iter); file; file = xar_file_next(iter)) {
-        char *pathCString;
-#if HAS_XAR_GET_SAFE_PATH
-        if (xar_get_safe_path != NULL) {
-            pathCString = xar_get_safe_path(file);
-        }
-        else
-#endif
-        {
-            pathCString = xar_get_path(file);
-        }
-        
-        if (pathCString == NULL) {
-            continue;
-        }
-        
-        NSString *path = @(pathCString);
-        if ([path.pathComponents containsObject:@".."]) {
-            xar_close(x);
-            
+    
+    if (![archive enumerateItems:^(const void * _Nonnull item, NSString * _Nonnull relativePath, SPUDeltaFileAttributes attributes, uint16_t permissions, BOOL *stop) {
+        if ([relativePath.pathComponents containsObject:@".."]) {
             if (error != NULL) {
-                *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Relative path '%@' contains '..' path component", path] }];
+                *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Relative path '%@' contains '..' path component", relativePath] }];
             }
-            return NO;
+            *stop = YES;
+            return;
         }
         
-        NSString *sourceFilePath = [source stringByAppendingPathComponent:path];
-        NSString *destinationFilePath = [destination stringByAppendingPathComponent:path];
+        NSString *sourceFilePath = [source stringByAppendingPathComponent:relativePath];
+        NSString *destinationFilePath = [destination stringByAppendingPathComponent:relativePath];
         {
             NSString *destinationParentDirectory = destinationFilePath.stringByDeletingLastPathComponent;
             NSDictionary<NSFileAttributeKey, id> *destinationParentDirectoryAttributes = [fileManager attributesOfItemAtPath:destinationParentDirectory error:NULL];
@@ -261,12 +173,11 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
                 // If it's inside a symlink, this is not good in any circumstance
                 NSString *fileType = destinationParentDirectoryAttributes[NSFileType];
                 if ([fileType isEqualToString:NSFileTypeSymbolicLink]) {
-                    xar_close(x);
-                    
                     if (error != NULL) {
                         *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Failed to create patch because '%@' cannot be a symbolic link.", destinationParentDirectory] }];
                     }
-                    return NO;
+                    *stop = YES;
+                    return;
                 }
             }
         }
@@ -277,86 +188,82 @@ BOOL applyBinaryDelta(NSString *source, NSString *destination, NSString *patchFi
         
         // Files that have no property set that we check for will get ignored
         // This is important because they aren't part of the delta, just part of the directory structure
-
-        const char *value;
-        if (xar_prop_get(file, DELETE_KEY, &value) == 0) {
+        
+        if ((attributes & SPUDeltaFileAttributesDelete) != 0) {
             if (!removeTree(destinationFilePath)) {
-                xar_close(x);
-                
                 if (verbose) {
                     fprintf(stderr, "\n");
                 }
                 if (error != NULL) {
-                    *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"%@: failed to remove %@", @DELETE_KEY, destination] }];
+                    *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"delete: failed to remove %@", destination] }];
                 }
-                return NO;
+                *stop = YES;
+                return;
             }
 
             removedFile = YES;
         }
 
-        if (xar_prop_get(file, BINARY_DELTA_KEY, &value) == 0) {
-            if (!applyBinaryDeltaToFile(x, file, sourceFilePath, destinationFilePath)) {
-                xar_close(x);
-                
+        if ((attributes & SPUDeltaFileAttributesBinaryDiff) != 0) {
+            if (!applyBinaryDeltaToFile(archive, item, sourceFilePath, destinationFilePath)) {
                 if (verbose) {
                     fprintf(stderr, "\n");
                 }
                 if (error != NULL) {
                     *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unable to patch %@ to destination %@", sourceFilePath, destinationFilePath] }];
                 }
-                return NO;
+                *stop = YES;
+                return;
             }
 
             if (verbose) {
-                fprintf(stderr, "\n🔨  %s %s", VERBOSE_PATCHED, [path fileSystemRepresentation]);
+                fprintf(stderr, "\n🔨  %s %s", VERBOSE_PATCHED, [relativePath fileSystemRepresentation]);
             }
-        } else if (xar_prop_get(file, EXTRACT_KEY, &value) == 0) { // extract and permission modifications don't coexist
-
-            if (xar_extract_tofile(x, file, [destinationFilePath fileSystemRepresentation]) != 0) {
-                xar_close(x);
-                
+        } else if ((attributes & SPUDeltaFileAttributesExtract) != 0) { // extract and permission modifications don't coexist
+            if (![archive extractItem:item destination:destinationFilePath]) {
                 if (verbose) {
                     fprintf(stderr, "\n");
                 }
                 if (error != NULL) {
                     *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteUnknownError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unable to extract file to %@", destinationFilePath] }];
                 }
-                return NO;
+                *stop = YES;
+                return;
             }
 
             if (verbose) {
                 if (fileExisted) {
-                    fprintf(stderr, "\n✏️  %s %s", VERBOSE_UPDATED, [path fileSystemRepresentation]);
+                    fprintf(stderr, "\n✏️  %s %s", VERBOSE_UPDATED, [relativePath fileSystemRepresentation]);
                 } else {
-                    fprintf(stderr, "\n✅  %s %s", VERBOSE_ADDED, [path fileSystemRepresentation]);
+                    fprintf(stderr, "\n✅  %s %s", VERBOSE_ADDED, [relativePath fileSystemRepresentation]);
                 }
             }
         } else if (verbose && removedFile) {
-            fprintf(stderr, "\n❌  %s %s", VERBOSE_DELETED, [path fileSystemRepresentation]);
+            fprintf(stderr, "\n❌  %s %s", VERBOSE_DELETED, [relativePath fileSystemRepresentation]);
         }
 
-        if (xar_prop_get(file, MODIFY_PERMISSIONS_KEY, &value) == 0) {
-            mode_t mode = (mode_t)[[NSString stringWithUTF8String:value] intValue];
+        if ((attributes & SPUDeltaFileAttributesModifyPermissions) != 0) {
+            mode_t mode = (mode_t)permissions;
             if (!modifyPermissions(destinationFilePath, mode)) {
-                xar_close(x);
-                
                 if (verbose) {
                     fprintf(stderr, "\n");
                 }
                 if (error != NULL) {
-                    *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteNoPermissionError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unable to modify permissions (%@) on file %@", @(value), destinationFilePath] }];
+                    *error = [NSError errorWithDomain:NSCocoaErrorDomain code:NSFileWriteNoPermissionError userInfo:@{ NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unable to modify permissions (%u) on file %@", mode, destinationFilePath] }];
                 }
-                return NO;
+                *stop = YES;
+                return;
             }
 
             if (verbose) {
-                fprintf(stderr, "\n👮  %s %s (0%o)", VERBOSE_MODIFIED, [path fileSystemRepresentation], mode);
+                fprintf(stderr, "\n👮  %s %s (0%o)", VERBOSE_MODIFIED, [relativePath fileSystemRepresentation], mode);
             }
         }
+    }]) {
+        return NO;
     }
     
-    xar_close(x);
+    [archive close];
 
     progressCallback(5/6.0);
